NT workaround, split, address review comments
